Changed so that not only integer formulas can be checked
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/mirjamsbranch@28321 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ doc/examples/example.c
@@ -1,6 +1,6 @@
 int main() {
-  int i = 0;
-  int a = 0;
+  double i = 0;
+  double a = 0;
 
   while (1) {
     if (i == 20) {
